refactor: test용 config 분리
