+++ src/test/java/chess/controller/LobbyControllerTest.java
@@ -7,40 +7,25 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
-import chess.controller.ChessGameControllerTest.HandlebarConfig;
+import chess.config.HandlebarConfig;
 import chess.domain.Score;
 import chess.domain.piece.Color;
 import chess.dto.ChessGameDto;
 import chess.dto.GameStatus;
 import chess.service.ChessGameService;
-import com.github.jknack.handlebars.springmvc.HandlebarsViewResolver;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Bean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
 
 @WebMvcTest(LobbyController.class)
 @ContextConfiguration(classes = HandlebarConfig.class)
 public class LobbyControllerTest {
-
-    @TestConfiguration
-    static class HandlebarConfig {
-
-        @Bean
-        public HandlebarsViewResolver handlebarsViewResolver() {
-            HandlebarsViewResolver resolver = new HandlebarsViewResolver();
-            resolver.setPrefix("classpath:/templates");
-            resolver.setSuffix(".hbs");
-            return resolver;
-        }
-    }
 
     @Autowired
     private MockMvc mockMvc;
